[L-02] Make Anvil fromVotingUnitDelta variable unsigned
Audit Feedback:
>In the `_moveVotingPower`  function of the Anvil contract, the integer sign of the `fromVotingUnitDelta` variable makes a distinction between the cases where voting units should be subtracted from, or added to, the from address.
However, this distinction is redundant since there is no case where `fromVotingUnitDelta` is a positive value. For this reason, there is also no need to cast `fromVotingUnitDelta` to a negative integer value. This variable could be declared as an unsigned integer and subtraction could always be performed when it is assigned a non-zero value.
>
>Consider declaring the `fromVotingUnitDelta` variable of the `_moveVotingPower` function as an unsigned integer in order to considerably simplify the code.

This change implements the suggestion in the last sentence above.