[audit suggestion] Removing maxPerAccount constraint from CollateralTokens
This was not mentioned in the audit report. This change comes from a conversation
during the most recent audit about the fact that `maxPerAccount` is not enforced
in all situations. While AF thinks there is some benefit in enforcing it in some
situations, the decision was made to remove it, as it is misleading.