Removing LetterOfCredit oracle update delay, allowance signatures, and making signatures EIP-712
1. Removing Oracle Update Delay
The previous version of the LetterOfCredit contract was careful to include
explicit delays in changes to its the price oracle as well as a delay to
change the delay. This was to provide the same features that a Governance
Timelock contract adds, namely, giving people who may be at risk due to
the change time to react to it.

It was written this way because the original thought was that LetterOfCredit
would be deployed prior to Anvil Governance. This is no longer necessary because
LetterOfCredit will be owned by Anvil's Governance contracts at the time
of deployment.

2. Allowance Signatures
This change also adds optional allowance signatures to be passed to
collateral-modifying functions so that the permission to increase collateral
usage can be provided and verified in a single transaction rather than requiring
multple.
This change required `ICollateral` to expose `getWithdrawalFeeBasisPoints()`
so that the `createLOC` function can accurately calculate the amount that is
expected in a Vault allowance modification signature.

3. EIP-712 Signatures
This change also makes LetterOfCredit signatures conform to EIP-712 so that
signers have a better idea of what they are signing.