Bug fixes and improvements for TBCP
Improvements:
- Add documentation to mapping state variables
- Replace i++ with ++i in loops to save gas
- Make constructor mark the singleton contract as initialized
- Prevent vault signature front-running

Bug Fixes:
1. Fix account reset ExitBalance calculation bug
2. Fix account reset determination logic

More details on (1):
Prior to this change, an account A to be reset could have its exit
balance from epoch X could be mixed together with other accounts
that have already processed the reset. The reset logic intends to
re-price units for a given token, so account A could theoretically
have way more units than other unstakers and remove well more than
their rightful amount when processing the unstake for exit X.

Note: this error only applies when all of the following conditions are met:
- The "attacker" needs to unstake right before a reset in the same epoch as the reset
- A reset just happened that significantly changes a unit's "worth"
- New stakers need to stake and unstake a large amount of value in the same
epoch that the reset occurred
- The attacker needs to wait until their unstake vests to process the unstake (i.e.
it cannot go down the reset route and instead must go down the unstake route)

More details on (2):
If an account calls unstake right after the pool is reset, the unstake can get
processed instead of just resetting, which can cause a number of downstream issues
for the account as well as that specific pool. The cause of this issue is not
properly determining if the account was reset in _releaseEligibleAccountTokens(...)
and returning instead whether the pool was reset. Both have been updated.